Switch Login and Registration forms to be use Django form base class
Fixes #6

This is a rewrite to the Django registration and login view code to make
it much easier to follow and maintain. This rewrite switches the logic
to use the Django forms base classes to reduce the amount of validation
and state maintenance that was previously done manually.
The majority of the validation is now done in UserRegistration\forms.py